requirements.txt
transformers
torch
pandas
scikit-learn
numpy
rouge_score
torchvision
wandb
pillow
kagglehub
sentencepiece
pytest-cov
pytest